Rename error & global_vars lib, separate compiler & main files
