+++ libs/compiler/codegen.c
@@ -1,10 +1,20 @@
-//
-//  codegen.c
-//
-//  Created by Andrey Terekhov on 3/10/15.
-//  Copyright (c) 2015 Andrey Terekhov. All rights reserved.
-//
-#include "global_vars.h"
+/*
+ *	Copyright 2015 Andrey Terekhov
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
+
+#include "global.h"
 #include <stdlib.h>
 
 extern void error(int err);
