+++ libs/compiler/global.h
@@ -1,10 +1,18 @@
-//
-//  global_vars.h
-//  RuC
-//
-//  Created by Andrey Terekhov on 03/06/14.
-//  Copyright (c) 2014 Andrey Terekhov. All rights reserved.
-/// Users/ant/Desktop/RuC/RuC/main.c
+/*
+ *	Copyright 2014 Andrey Terekhov
+ *
+ *	Licensed under the Apache License, Version 2.0 (the "License");
+ *	you may not use this file except in compliance with the License.
+ *	You may obtain a copy of the License at
+ *
+ *		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *	Unless required by applicable law or agreed to in writing, software
+ *	distributed under the License is distributed on an "AS IS" BASIS,
+ *	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *	See the License for the specific language governing permissions and
+ *	limitations under the License.
+ */
 
 #include <stdio.h>
